+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-Current dev
+0.4.0 (2020-03-14)
 ===========
 * Added new statistic "jump-align" which counts the percentage of instructions that jump to unaligned addresses
 * Tried to improve documentation
+* Changed `base` to use non-cyclic mode by default (meaning that firmware doesn't wrap on 0x10000 when searching for alignment), old cyclic mode can be accessed with `--cyclic`
+* Added `--dump` to base, which dumps the likeliness of every possible offset
+* Added rudimentary config for libfind libraries and stat mode
 
 0.3.0 (2019-10-17)
 ==================